requirements.txt
flet
pypdf
PyMuPDF
Pillow
plyer
